Refactor: comment service 디렉토리 추가
