+++ apps/api/src/comment/comment-api.module.ts
@@ -2,9 +2,9 @@ import { ArticleModule } from '@api/article/article.module';
 import { CategoryModule } from '@api/category/category.module';
 import { CommentApiController } from '@api/comment/comment-api.controller';
 import { CommentModule } from '@api/comment/comment.module';
-import { CreateCommentApiService } from '@api/comment/create-comment-api.service';
-import { RemoveCommentApiService } from '@api/comment/remove-comment-api.service';
-import { UpdateCommentApiService } from '@api/comment/update-comment-api.service';
+import { CreateCommentApiService } from '@api/comment/services/create-comment-api.service';
+import { RemoveCommentApiService } from '@api/comment/services/remove-comment-api.service';
+import { UpdateCommentApiService } from '@api/comment/services/update-comment-api.service';
 import { NotificationModule } from '@api/notification/notification.module';
 import { Module } from '@nestjs/common';
