+++ apps/api/src/comment/comment.module.ts
@@ -1,7 +1,7 @@
 import { CommentRepository } from '@api/comment/repositories/comment.repository';
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { CommentService } from './comment.service';
+import { CommentService } from './services/comment.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([CommentRepository])],
