+++ apps/api/src/comment/services/update-comment-api.service.ts
@@ -1,4 +1,4 @@
-import { CommentService } from '@api/comment/comment.service';
+import { CommentService } from '@api/comment/services/comment.service';
 import { Injectable } from '@nestjs/common';
 
 @Injectable()
